Fix: Mv mock related stuff into sperate file to prevent bloat
