+++ test/mocks.dart
@@ -1,0 +1,21 @@
+import 'package:decentproof/features/metadata/logic/LocationService.dart';
+import 'package:decentproof/features/metadata/logic/MetaDataPermissionService.dart';
+import 'package:decentproof/shared/Integrety/ApiKeyRequestService.dart';
+import 'package:decentproof/shared/Integrety/AppCheck.dart';
+import 'package:decentproof/shared/Integrety/SecureStorageWrapper.dart';
+import 'package:mockito/annotations.dart';
+
+///
+///This is just a placeholder file for the genertion of mocks
+///To prevent dupplication of generated code, all classes can be registered here for generation.
+///
+
+@GenerateNiceMocks([
+  MockSpec<AppCheck>(),
+  MockSpec<ApiKeyRequestService>(),
+  MockSpec<SecureStorageService>(),
+  MockSpec<LocationService>(),
+  MockSpec<MetaDataPermissionService>(),
+])
+// ignore: unused_element
+void main() {}
