+++ test/mocks.mocks.dart
@@ -1,15 +1,18 @@
 // Mocks generated by Mockito 5.4.2 from annotations
-// in decentproof/test/metadata/bloc/LocationWarningBloc_test.dart.
+// in decentproof/test/mocks.dart.
 // Do not manually edit this file.
 
 // ignore_for_file: no_leading_underscores_for_library_prefixes
 import 'dart:async' as _i4;
 
 import 'package:decentproof/features/metadata/logic/LocationService.dart'
-    as _i3;
+    as _i7;
 import 'package:decentproof/features/metadata/logic/MetaDataPermissionService.dart'
-    as _i5;
+    as _i8;
 import 'package:decentproof/features/metadata/models/LocationModel.dart' as _i2;
+import 'package:decentproof/shared/Integrety/ApiKeyRequestService.dart' as _i5;
+import 'package:decentproof/shared/Integrety/AppCheck.dart' as _i3;
+import 'package:decentproof/shared/Integrety/SecureStorageWrapper.dart' as _i6;
 import 'package:mockito/mockito.dart' as _i1;
 
 // ignore_for_file: type=lint
@@ -33,10 +36,111 @@ class _FakeLocationModel_0 extends _i1.SmartFake implements _i2.LocationModel {
         );
 }
 
+/// A class which mocks [AppCheck].
+///
+/// See the documentation for Mockito's code generation for more information.
+class MockAppCheck extends _i1.Mock implements _i3.AppCheck {
+  @override
+  _i4.Future<String> getIntegrityToken() => (super.noSuchMethod(
+        Invocation.method(
+          #getIntegrityToken,
+          [],
+        ),
+        returnValue: _i4.Future<String>.value(''),
+        returnValueForMissingStub: _i4.Future<String>.value(''),
+      ) as _i4.Future<String>);
+}
+
+/// A class which mocks [ApiKeyRequestService].
+///
+/// See the documentation for Mockito's code generation for more information.
+class MockApiKeyRequestService extends _i1.Mock
+    implements _i5.ApiKeyRequestService {
+  @override
+  _i4.Future<bool> checkForNewApiKey(String? apiKey) => (super.noSuchMethod(
+        Invocation.method(
+          #checkForNewApiKey,
+          [apiKey],
+        ),
+        returnValue: _i4.Future<bool>.value(false),
+        returnValueForMissingStub: _i4.Future<bool>.value(false),
+      ) as _i4.Future<bool>);
+  @override
+  _i4.Future<String> getNewNewKey(String? token) => (super.noSuchMethod(
+        Invocation.method(
+          #getNewNewKey,
+          [token],
+        ),
+        returnValue: _i4.Future<String>.value(''),
+        returnValueForMissingStub: _i4.Future<String>.value(''),
+      ) as _i4.Future<String>);
+}
+
+/// A class which mocks [SecureStorageService].
+///
+/// See the documentation for Mockito's code generation for more information.
+class MockSecureStorageService extends _i1.Mock
+    implements _i6.SecureStorageService {
+  @override
+  _i4.Future<void> saveApiKey(String? key) => (super.noSuchMethod(
+        Invocation.method(
+          #saveApiKey,
+          [key],
+        ),
+        returnValue: _i4.Future<void>.value(),
+        returnValueForMissingStub: _i4.Future<void>.value(),
+      ) as _i4.Future<void>);
+  @override
+  _i4.Future<String?> retriveApiKey() => (super.noSuchMethod(
+        Invocation.method(
+          #retriveApiKey,
+          [],
+        ),
+        returnValue: _i4.Future<String?>.value(),
+        returnValueForMissingStub: _i4.Future<String?>.value(),
+      ) as _i4.Future<String?>);
+  @override
+  _i4.Future<String?> retriveEmail() => (super.noSuchMethod(
+        Invocation.method(
+          #retriveEmail,
+          [],
+        ),
+        returnValue: _i4.Future<String?>.value(),
+        returnValueForMissingStub: _i4.Future<String?>.value(),
+      ) as _i4.Future<String?>);
+  @override
+  _i4.Future<void> saveEmail(String? email) => (super.noSuchMethod(
+        Invocation.method(
+          #saveEmail,
+          [email],
+        ),
+        returnValue: _i4.Future<void>.value(),
+        returnValueForMissingStub: _i4.Future<void>.value(),
+      ) as _i4.Future<void>);
+  @override
+  _i4.Future<String?> retriveSecret() => (super.noSuchMethod(
+        Invocation.method(
+          #retriveSecret,
+          [],
+        ),
+        returnValue: _i4.Future<String?>.value(),
+        returnValueForMissingStub: _i4.Future<String?>.value(),
+      ) as _i4.Future<String?>);
+  @override
+  _i4.Future<void> saveSecret(String? secret) => (super.noSuchMethod(
+        Invocation.method(
+          #saveSecret,
+          [secret],
+        ),
+        returnValue: _i4.Future<void>.value(),
+        returnValueForMissingStub: _i4.Future<void>.value(),
+      ) as _i4.Future<void>);
+}
+
 /// A class which mocks [LocationService].
 ///
 /// See the documentation for Mockito's code generation for more information.
-class MockLocationService extends _i1.Mock implements _i3.LocationService {
+class MockLocationService extends _i1.Mock implements _i7.LocationService {
   @override
   _i4.Future<bool> getPermission() => (super.noSuchMethod(
         Invocation.method(
@@ -92,7 +196,7 @@ class MockLocationService extends _i1.Mock implements _i3.LocationService {
 ///
 /// See the documentation for Mockito's code generation for more information.
 class MockMetaDataPermissionService extends _i1.Mock
-    implements _i5.MetaDataPermissionService {
+    implements _i8.MetaDataPermissionService {
   @override
   _i4.Future<void> init() => (super.noSuchMethod(
         Invocation.method(
