--- test/settings/bloc/SettingsBloc_test.mocks.dart
@@ -1,203 +1,0 @@
-// Mocks generated by Mockito 5.4.2 from annotations
-// in decentproof/test/settings/bloc/SettingsBloc_test.dart.
-// Do not manually edit this file.
-
-// ignore_for_file: no_leading_underscores_for_library_prefixes
-import 'dart:async' as _i4;
-
-import 'package:decentproof/features/metadata/logic/LocationService.dart'
-    as _i3;
-import 'package:decentproof/features/metadata/logic/MetaDataPermissionService.dart'
-    as _i5;
-import 'package:decentproof/features/metadata/models/LocationModel.dart' as _i2;
-import 'package:decentproof/shared/Integrety/SecureStorageWrapper.dart' as _i6;
-import 'package:mockito/mockito.dart' as _i1;
-
-// ignore_for_file: type=lint
-// ignore_for_file: avoid_redundant_argument_values
-// ignore_for_file: avoid_setters_without_getters
-// ignore_for_file: comment_references
-// ignore_for_file: implementation_imports
-// ignore_for_file: invalid_use_of_visible_for_testing_member
-// ignore_for_file: prefer_const_constructors
-// ignore_for_file: unnecessary_parenthesis
-// ignore_for_file: camel_case_types
-// ignore_for_file: subtype_of_sealed_class
-
-class _FakeLocationModel_0 extends _i1.SmartFake implements _i2.LocationModel {
-  _FakeLocationModel_0(
-    Object parent,
-    Invocation parentInvocation,
-  ) : super(
-          parent,
-          parentInvocation,
-        );
-}
-
-/// A class which mocks [LocationService].
-///
-/// See the documentation for Mockito's code generation for more information.
-class MockLocationService extends _i1.Mock implements _i3.LocationService {
-  @override
-  _i4.Future<bool> getPermission() => (super.noSuchMethod(
-        Invocation.method(
-          #getPermission,
-          [],
-        ),
-        returnValue: _i4.Future<bool>.value(false),
-        returnValueForMissingStub: _i4.Future<bool>.value(false),
-      ) as _i4.Future<bool>);
-  @override
-  _i4.Future<bool> hasPermission() => (super.noSuchMethod(
-        Invocation.method(
-          #hasPermission,
-          [],
-        ),
-        returnValue: _i4.Future<bool>.value(false),
-        returnValueForMissingStub: _i4.Future<bool>.value(false),
-      ) as _i4.Future<bool>);
-  @override
-  _i4.Future<_i2.LocationModel> requestLocation() => (super.noSuchMethod(
-        Invocation.method(
-          #requestLocation,
-          [],
-        ),
-        returnValue: _i4.Future<_i2.LocationModel>.value(_FakeLocationModel_0(
-          this,
-          Invocation.method(
-            #requestLocation,
-            [],
-          ),
-        )),
-        returnValueForMissingStub:
-            _i4.Future<_i2.LocationModel>.value(_FakeLocationModel_0(
-          this,
-          Invocation.method(
-            #requestLocation,
-            [],
-          ),
-        )),
-      ) as _i4.Future<_i2.LocationModel>);
-  @override
-  _i4.Future<bool> serviceEnabled() => (super.noSuchMethod(
-        Invocation.method(
-          #serviceEnabled,
-          [],
-        ),
-        returnValue: _i4.Future<bool>.value(false),
-        returnValueForMissingStub: _i4.Future<bool>.value(false),
-      ) as _i4.Future<bool>);
-}
-
-/// A class which mocks [MetaDataPermissionService].
-///
-/// See the documentation for Mockito's code generation for more information.
-class MockMetaDataPermissionService extends _i1.Mock
-    implements _i5.MetaDataPermissionService {
-  @override
-  _i4.Future<void> init() => (super.noSuchMethod(
-        Invocation.method(
-          #init,
-          [],
-        ),
-        returnValue: _i4.Future<void>.value(),
-        returnValueForMissingStub: _i4.Future<void>.value(),
-      ) as _i4.Future<void>);
-  @override
-  _i4.Future<void> allowLocationEmbedding(bool? allow) => (super.noSuchMethod(
-        Invocation.method(
-          #allowLocationEmbedding,
-          [allow],
-        ),
-        returnValue: _i4.Future<void>.value(),
-        returnValueForMissingStub: _i4.Future<void>.value(),
-      ) as _i4.Future<void>);
-  @override
-  _i4.Future<void> allowSecretEmbedding(bool? allow) => (super.noSuchMethod(
-        Invocation.method(
-          #allowSecretEmbedding,
-          [allow],
-        ),
-        returnValue: _i4.Future<void>.value(),
-        returnValueForMissingStub: _i4.Future<void>.value(),
-      ) as _i4.Future<void>);
-  @override
-  bool shouldEmbedLocation() => (super.noSuchMethod(
-        Invocation.method(
-          #shouldEmbedLocation,
-          [],
-        ),
-        returnValue: false,
-        returnValueForMissingStub: false,
-      ) as bool);
-  @override
-  bool shouldEmbedSecret() => (super.noSuchMethod(
-        Invocation.method(
-          #shouldEmbedSecret,
-          [],
-        ),
-        returnValue: false,
-        returnValueForMissingStub: false,
-      ) as bool);
-}
-
-/// A class which mocks [SecureStorageService].
-///
-/// See the documentation for Mockito's code generation for more information.
-class MockSecureStorageService extends _i1.Mock
-    implements _i6.SecureStorageService {
-  @override
-  _i4.Future<void> saveApiKey(String? key) => (super.noSuchMethod(
-        Invocation.method(
-          #saveApiKey,
-          [key],
-        ),
-        returnValue: _i4.Future<void>.value(),
-        returnValueForMissingStub: _i4.Future<void>.value(),
-      ) as _i4.Future<void>);
-  @override
-  _i4.Future<String?> retriveApiKey() => (super.noSuchMethod(
-        Invocation.method(
-          #retriveApiKey,
-          [],
-        ),
-        returnValue: _i4.Future<String?>.value(),
-        returnValueForMissingStub: _i4.Future<String?>.value(),
-      ) as _i4.Future<String?>);
-  @override
-  _i4.Future<String?> retriveEmail() => (super.noSuchMethod(
-        Invocation.method(
-          #retriveEmail,
-          [],
-        ),
-        returnValue: _i4.Future<String?>.value(),
-        returnValueForMissingStub: _i4.Future<String?>.value(),
-      ) as _i4.Future<String?>);
-  @override
-  _i4.Future<void> saveEmail(String? email) => (super.noSuchMethod(
-        Invocation.method(
-          #saveEmail,
-          [email],
-        ),
-        returnValue: _i4.Future<void>.value(),
-        returnValueForMissingStub: _i4.Future<void>.value(),
-      ) as _i4.Future<void>);
-  @override
-  _i4.Future<String?> retriveSecret() => (super.noSuchMethod(
-        Invocation.method(
-          #retriveSecret,
-          [],
-        ),
-        returnValue: _i4.Future<String?>.value(),
-        returnValueForMissingStub: _i4.Future<String?>.value(),
-      ) as _i4.Future<String?>);
-  @override
-  _i4.Future<void> saveSecret(String? secret) => (super.noSuchMethod(
-        Invocation.method(
-          #saveSecret,
-          [secret],
-        ),
-        returnValue: _i4.Future<void>.value(),
-        returnValueForMissingStub: _i4.Future<void>.value(),
-      ) as _i4.Future<void>);
-}
