--- test/shared/util/RequestUtil_test.mocks.dart
@@ -1,123 +1,0 @@
-// Mocks generated by Mockito 5.4.2 from annotations
-// in decentproof/test/shared/util/RequestUtil_test.dart.
-// Do not manually edit this file.
-
-// ignore_for_file: no_leading_underscores_for_library_prefixes
-import 'dart:async' as _i3;
-
-import 'package:decentproof/shared/Integrety/ApiKeyRequestService.dart' as _i4;
-import 'package:decentproof/shared/Integrety/AppCheck.dart' as _i2;
-import 'package:decentproof/shared/Integrety/SecureStorageWrapper.dart' as _i5;
-import 'package:mockito/mockito.dart' as _i1;
-
-// ignore_for_file: type=lint
-// ignore_for_file: avoid_redundant_argument_values
-// ignore_for_file: avoid_setters_without_getters
-// ignore_for_file: comment_references
-// ignore_for_file: implementation_imports
-// ignore_for_file: invalid_use_of_visible_for_testing_member
-// ignore_for_file: prefer_const_constructors
-// ignore_for_file: unnecessary_parenthesis
-// ignore_for_file: camel_case_types
-// ignore_for_file: subtype_of_sealed_class
-
-/// A class which mocks [AppCheck].
-///
-/// See the documentation for Mockito's code generation for more information.
-class MockAppCheck extends _i1.Mock implements _i2.AppCheck {
-  @override
-  _i3.Future<String> getIntegrityToken() => (super.noSuchMethod(
-        Invocation.method(
-          #getIntegrityToken,
-          [],
-        ),
-        returnValue: _i3.Future<String>.value(''),
-        returnValueForMissingStub: _i3.Future<String>.value(''),
-      ) as _i3.Future<String>);
-}
-
-/// A class which mocks [ApiKeyRequestService].
-///
-/// See the documentation for Mockito's code generation for more information.
-class MockApiKeyRequestService extends _i1.Mock
-    implements _i4.ApiKeyRequestService {
-  @override
-  _i3.Future<bool> checkForNewApiKey(String? apiKey) => (super.noSuchMethod(
-        Invocation.method(
-          #checkForNewApiKey,
-          [apiKey],
-        ),
-        returnValue: _i3.Future<bool>.value(false),
-        returnValueForMissingStub: _i3.Future<bool>.value(false),
-      ) as _i3.Future<bool>);
-  @override
-  _i3.Future<String> getNewNewKey(String? token) => (super.noSuchMethod(
-        Invocation.method(
-          #getNewNewKey,
-          [token],
-        ),
-        returnValue: _i3.Future<String>.value(''),
-        returnValueForMissingStub: _i3.Future<String>.value(''),
-      ) as _i3.Future<String>);
-}
-
-/// A class which mocks [SecureStorageService].
-///
-/// See the documentation for Mockito's code generation for more information.
-class MockSecureStorageService extends _i1.Mock
-    implements _i5.SecureStorageService {
-  @override
-  _i3.Future<void> saveApiKey(String? key) => (super.noSuchMethod(
-        Invocation.method(
-          #saveApiKey,
-          [key],
-        ),
-        returnValue: _i3.Future<void>.value(),
-        returnValueForMissingStub: _i3.Future<void>.value(),
-      ) as _i3.Future<void>);
-  @override
-  _i3.Future<String?> retriveApiKey() => (super.noSuchMethod(
-        Invocation.method(
-          #retriveApiKey,
-          [],
-        ),
-        returnValue: _i3.Future<String?>.value(),
-        returnValueForMissingStub: _i3.Future<String?>.value(),
-      ) as _i3.Future<String?>);
-  @override
-  _i3.Future<String?> retriveEmail() => (super.noSuchMethod(
-        Invocation.method(
-          #retriveEmail,
-          [],
-        ),
-        returnValue: _i3.Future<String?>.value(),
-        returnValueForMissingStub: _i3.Future<String?>.value(),
-      ) as _i3.Future<String?>);
-  @override
-  _i3.Future<void> saveEmail(String? email) => (super.noSuchMethod(
-        Invocation.method(
-          #saveEmail,
-          [email],
-        ),
-        returnValue: _i3.Future<void>.value(),
-        returnValueForMissingStub: _i3.Future<void>.value(),
-      ) as _i3.Future<void>);
-  @override
-  _i3.Future<String?> retriveSecret() => (super.noSuchMethod(
-        Invocation.method(
-          #retriveSecret,
-          [],
-        ),
-        returnValue: _i3.Future<String?>.value(),
-        returnValueForMissingStub: _i3.Future<String?>.value(),
-      ) as _i3.Future<String?>);
-  @override
-  _i3.Future<void> saveSecret(String? secret) => (super.noSuchMethod(
-        Invocation.method(
-          #saveSecret,
-          [secret],
-        ),
-        returnValue: _i3.Future<void>.value(),
-        returnValueForMissingStub: _i3.Future<void>.value(),
-      ) as _i3.Future<void>);
-}
